% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_biplot_arrows}
\alias{add_biplot_arrows}
\title{Adds a biplot arrow layer to an existing plot}
\usage{
add_biplot_arrows(object, plt, vars, colour = "navy")
}
\arguments{
\item{object}{a fortified ordination object.}

\item{plt}{a ggplot object.}

\item{vars}{character; length 2 vector of dimension names.}
}
\description{
Adds a biplot arrow layer to an existing plot
}
