% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_spp_site_scores}
\alias{add_spp_site_scores}
\title{Adds species and site score layers to an existing plot}
\usage{
add_spp_site_scores(object, plt, vars, geom, draw_list, arrows)
}
\arguments{
\item{object}{an ordination object.}

\item{plt}{a ggplot object.}

\item{vars}{character; length 2 vector of dimension names.}

\item{geom}{character; vector of length 1 or 2 indicating which geoms will
be used ofr the species or site scores.}

\item{draw_list}{logical; vector of types of scores indicating which are
available and requested for plotting.}

\item{arrows}{logical; length 1 vector indicating if species scores should
be drawn using arrows.}
}
\description{
Adds species and site score layers to an existing plot
}
