% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arrowMul}
\alias{arrowMul}
\title{Scale Vectors to Data}
\usage{
arrowMul(arrows, data, at = c(0, 0), fill = 0.75)
}
\arguments{
\item{arrows}{a two-column matrix-like object containing coordinates for the arrows/vectors on x and y axes.}

\item{data}{a two-column matrix-like object containing coordinates of the data on the x and y axes.}

\item{at}{numeric vector of length 2; location of the origin of the arrows.}

\item{fill}{numeric; what proportion of the range of the data to fill}
}
\value{
a numeric multiplier that will scale the arrows
}
\description{
Scale vector arrows to \code{fill} proportion of the data.
}
\author{
Gavin L. Simpson
}
