% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.anosim.R
\name{autoplot.anosim}
\alias{autoplot.anosim}
\title{ggplot-based plot for objects of class \code{"anosim"}}
\usage{
\method{autoplot}{anosim}(object, notch = TRUE, varwidth = TRUE,
  subtitle = TRUE, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{"anosim"}, the result of a call to \code{\link[vegan]{anosim}}.}

\item{notch}{logical; make notched (default) or standard box plot?}

\item{varwidth}{logical; make box width proportional to to the square-roots of the number of observations in the groups (default) or standart box plot?}

\item{subtitle}{logical; should subtitle with R and p values be added.}

\item{xlab}{character; label for the x-axis.}

\item{ylab}{character; label for the y-axis.}

\item{...}{additional arguments passed to other methods.}
}
\value{
A ggplot object.
}
\description{
Produces a multi-layer ggplot object representing the output of objects produced by \code{\link[vegan]{anosim}}.
}
\examples{

data(dune)
data(dune.env)
dune.dist <- vegdist(dune)
attach(dune.env)
dune.ano <- anosim(dune.dist, Management)

autoplot(dune.ano)
}
\author{
Didzis Elferts
}
