% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.fisherfit.R
\name{autoplot.fisherfit}
\alias{autoplot.fisherfit}
\title{Plot Fisher's log-series}
\usage{
\method{autoplot}{fisherfit}(object, xlab = "Rank", ylab = "Abundance",
  bar.col = NA, bar.fill = "grey35", line.col = "red", size = 1, ...)
}
\arguments{
\item{object}{an object of class \code{\link{fisherfit}}.}

\item{xlab}{character; label for the x axis.}

\item{ylab}{character; label for the y axis.}

\item{bar.col}{colour for the bar outlines. The default, \code{NA}, does not draw outlines around bars.}

\item{bar.fill}{fill colour for the bars.}

\item{line.col}{colour for Fisher's log-series curve.}

\item{size}{numeric; size aesthetic for the log-series curve.}

\item{...}{additional arguments passed to other methods.}
}
\value{
A ggplot object.
}
\description{
Draws a bar plot of species rank abundance with Fisher's log-series superimposed.
}
\examples{

data(BCI)
mod <- fisherfit(BCI[5,])
autoplot(mod)
}
\author{
Gavin L. Simpson
}
