% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.permat.R
\name{autoplot.permat}
\alias{autoplot.permat}
\title{ggplot-based plot for objects of class \code{permat}}
\usage{
\method{autoplot}{permat}(object, col = c("red", "blue"), type = "bray",
  lowess = TRUE, xlab = "Runs", ylab = NULL,
  title = "Results of matrix permutation algorithms", subtitle = NULL,
  caption = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{permat}, the result of a call to \code{\link[vegan]{permatswap}}.}

\item{col}{character; colors of Dissimilarity line and lowess line.}

\item{type}{character; type of plot to be displayed: "bray" for Bray-Curtis dissimilarities, "chisq" for Chi-squared values.}

\item{lowess}{logical; whether a locally weighted regression curve should be drawn.}

\item{xlab}{character; label for the x axis.}

\item{ylab}{character; label for the y axis.}

\item{title}{character; title for the plot.}

\item{subtitle}{character; subtitle for the plot.}

\item{caption}{character; caption for the plot.}

\item{...}{additional arguments passed to other methods.}
}
\value{
A ggplot object.
}
\description{
Produces a multi-layer ggplot object representing the output of objects produced by \code{\link[vegan]{permatswap}}.
}
\examples{

data(BCI)
perm <- permatswap(BCI, "quasiswap", times=19)

autoplot(perm)
}
\author{
Didzis Elferts
}
