% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.rad.R
\name{autoplot.rad}
\alias{autoplot.rad}
\title{ggplot-based plot for objects of class \code{"rad"}}
\usage{
\method{autoplot}{rad}(object, log = "y", xlab = "Rank",
  ylab = "Abundance", title = "Species abundance against species rank",
  subtitle = NULL, caption = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{rad}.}

\item{log}{character; use logarithmic scale for "x" or "y", "xy" makes both axis logarthmic and "none" use original scale.}

\item{xlab}{character; label for the x axis.}

\item{ylab}{character; label for the y axis.}

\item{title}{character; title for the plot.}

\item{subtitle}{character; subtitle for the plot.}

\item{caption}{character; caption for the plot.}

\item{...}{additional arguments passed to other methods.}
}
\value{
A ggplot object.
}
\description{
Produces a multi-layer ggplot object representing the output of object \code{rad} produced by \code{\link[vegan]{radfit}}, \code{\link[vegan]{rad.null}}, \code{\link[vegan]{rad.preempt}}, \code{\link[vegan]{rad.lognormal}}, \code{\link[vegan]{rad.zipf}} or \code{\link[vegan]{rad.zipfbrot}}.
}
\examples{

data(BCI)
mod <- rad.lognormal(BCI[5,])

autoplot(mod$y)
}
\author{
Didzis Elferts
}
