% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.radfit.R
\name{autoplot.radfit}
\alias{autoplot.radfit}
\alias{autoplot.radfit.frame}
\alias{autoplot.radline}
\alias{autoplot.rad}
\alias{autoplot.rad.frame}
\title{Autoplot graphics for vegan rank-abundance models}
\usage{
\method{autoplot}{radfit}(
  object,
  facet = TRUE,
  point.params = list(),
  line.params = list(),
  ...
)

\method{autoplot}{radfit.frame}(object, point.params = list(), line.params = list(), ...)

\method{autoplot}{radline}(object, point.params = list(), line.params = list(), ...)

\method{autoplot}{rad}(object, point.params = list(), line.params = list(), ...)

\method{autoplot}{rad.frame}(object, point.params = list(), highlight = NULL, ...)
}
\arguments{
\item{object}{Result object from \code{\link[vegan]{radfit}}.}

\item{facet}{Draw each fitted model to a separate facet or (if
\code{FALSE}) all fitted lines to a single graph.}

\item{point.params, line.params}{Parameters to modify points or
lines (passed to \code{\link[ggplot2]{geom_point}} and
\code{\link[ggplot2]{geom_line}}).}

\item{...}{Additional arguments passed to the functions.}

\item{highlight}{Names of species that should be highlighted as
coloured points.}
}
\description{
\pkg{ggplot2} graphics for Rank-Abundance Distribution models
fitted with \pkg{vegan} functions \code{\link[vegan:radfit]{vegan::radfit()}} or produced
with \code{\link[vegan:radfit]{vegan::as.rad()}}.
}
\details{
The \code{autoplot} function draws graphics which are \pkg{ggplot2}
alternatives for \CRANpkg{lattice} graphics in \pkg{vegan}. In
addition, there are functions for \code{\link[vegan:radfit]{vegan::as.rad()}}
results which do not have dedicated graphics in\pkg{vegan}.
}
\examples{
library(vegan)
library(ggplot2)
data(mite)
m1 <- radfit(mite[1,])
## With logarithmic y-axis (default) Pre-emption model is a line
autoplot(m1) + labs(title="log-Abundance: Pre-emption model is a line")
## With log-log scale, Zipf model is a line
autoplot(m1) + scale_x_log10() +
   labs(title="log-log Scale: Zipf model is a line")
## Show only the best model
autoplot(m1, pick = "AIC")
## Show selected models in one frame
autoplot(m1, pick = c("Z","M","L"), facet=FALSE)
## plot best models for several sites
m <- radfit(mite[1:12,])
autoplot(m) + labs(title="Model Selection AIC (Default)")
## use BIC and reoreder sites by their diversity
autoplot(m, pick="BIC", order.by = diversity(mite[1:12,])) +
   labs(title="Model Selection BIC, Ordered by Increasing Diversity")
## Plot RAD models without fits highlighting most abundant species in the
## whole data.
m0 <- as.rad(mite[1:12,])
dominants <- names(sort(colSums(mite), decreasing=TRUE))[1:6]
autoplot(m0, highlight = dominants)
}
