% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.renyiaccum.R
\name{autoplot.renyiaccum}
\alias{autoplot.renyiaccum}
\title{ggplot-based plot for objects of class \code{"renyiaccum"}}
\usage{
\method{autoplot}{renyiaccum}(object, facet = TRUE, ribbon = facet,
  ncol = NULL, ribbon.alpha = 0.3, xlab = "Sites", ylab = "Diversity",
  title = "Diversity accumulation curves", subtitle = NULL,
  caption = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{renyiaccum}, the result of a call to \code{\link[vegan]{renyiaccum}}.}

\item{facet}{logical; should diversity values be shown in separate facets?}

\item{ribbon}{logical; show the quantile-based uncertainty interval? Uses \code{\link{geom_ribbon}} for plot.}

\item{ncol}{numeric; if facetting the plot, how many columns to use.}

\item{ribbon.alpha}{numeric; alpha transparency used for the uncertainty interval. Passed to the \code{alpha} aesthetic of \code{\link[ggplot2]{geom_ribbon}}.}

\item{xlab}{character; label for the x axis.}

\item{ylab}{character; label for the y axis.}

\item{title}{character; title for the plot.}

\item{subtitle}{character; subtitle for the plot.}

\item{caption}{character; caption for the plot.}

\item{...}{additional arguments passed to other methods.}
}
\value{
A ggplot object.
}
\description{
Produces a multi-layer ggplot object representing the output of objects produced by \code{\link[vegan]{renyiaccum}}.
}
\examples{

data(BCI)
i <- sample(nrow(BCI), 12)
mod <- renyiaccum(BCI[i,])

autoplot(mod)
## Turn off facetting; turns off ribbon too
autoplot(mod, facet = FALSE)
}
\author{
Didzis Elferts & Gavin L. Simpson
}
