% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.wcmdscale.R
\name{autoplot.wcmdscale}
\alias{autoplot.wcmdscale}
\title{ggplot-based plot for objects of class \code{"wcmdscale"}}
\usage{
\method{autoplot}{wcmdscale}(object, geom = c("point", "text"), ylab, xlab,
  ...)
}
\arguments{
\item{object}{an object of class \code{"wcmdscale"}, the result of a call
to \code{\link[vegan]{wcmdscale}} with eig = TRUE.}

\item{geom}{character; which geoms to use for the layer.}

\item{ylab}{character; label for the y-axis}

\item{xlab}{character; label for the x-axis}

\item{...}{Additional arguments passed to \code{\link{fortify.wcmdscale}}.}
}
\value{
Returns a ggplot object.
}
\description{
Produces a  ggplot object representing the output of
objects produced by \code{\link[vegan]{wcmdscale}}.
}
\details{
TODO
}
\examples{
## load example data
data(dune)
d <- vegdist(decostand(dune, method = 'pa'), method = 'jaccard')
ord <- wcmdscale(d, eig = TRUE)
autoplot(ord)
 autoplot(ord, geom = 'text')
}
\author{
Eduard Szöcs eduardszoecs@gmail.com
}

