% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_user_layers}
\alias{check_user_layers}
\title{Check user-supplied layers against list of valid layers}
\usage{
check_user_layers(user, valid, message = FALSE)
}
\arguments{
\item{user}{character; vector of user supplied layer names.}

\item{valid}{character; vector of valid layer names.}

\item{message}{logical; should a message be raised in the case of invalid
user-supplied layer names.}
}
\description{
Check user-supplied layers against list of valid layers
}
