% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.permat.R
\name{fortify.permat}
\alias{fortify.permat}
\title{Fortify matrix permutation algorithms for presence-absence and count data results}
\usage{
\method{fortify}{permat}(model, data, ...)
}
\arguments{
\item{model}{an object of class \code{permat}.}

\item{data}{original data set. Currently ignored.}

\item{...}{additional arguments for other methods. Currently ignored.}
}
\value{
A data frame with columns \code{Runs}, \code{Dissimilarity} and \code{Type}, showing number of runs, dissimilarity index and type of dissimilarity values.
}
\description{
Prepares a fortified version of results from \code{permat} objects.
}
\examples{
data(BCI)
perm <- permatswap(BCI, "quasiswap", times=19)
df <- fortify(perm)

ggplot(df, aes(x = Runs, y = Dissimilarity)) + 
   geom_line() +
   facet_wrap(~Type, ncol = 2) 
   
}
\author{
Didzis Elferts
}
