% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.permustats.R
\name{fortify.permustats}
\alias{fortify.permustats}
\title{Fortify permutation statistics}
\usage{
\method{fortify}{permustats}(model, data, scale = FALSE, ...)
}
\arguments{
\item{model}{an object of created by \code{\link[vegan:permustats]{vegan::permustats()}}.}

\item{data}{original data set. Currently ignored.}

\item{scale}{logical; return standardized effect sizes (SES)?}

\item{...}{Other parameters passed to functions (ignored).}
}
\value{
A tibble with columns \code{permutations}, and \code{terms} containing the
values of tests statistics under the null hypothesis, and a factor
labelling the permutation, respectively.
}
\description{
Fortify permutation statistics
}
