% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.rad.R
\name{fortify.rad}
\alias{fortify.rad}
\title{Fortify \code{rad} object}
\usage{
\method{fortify}{rad}(model, data, ...)
}
\arguments{
\item{model}{an object of class \code{rad}.}

\item{data}{original data set. Currently ignored.}

\item{...}{additional arguments for other methods. Currently ignored.}
}
\value{
A data frame with columns \code{Species}, \code{Rank} and \code{Abundance}, showing species name, their rank and abundance values.
}
\description{
Prepares  a fortified version of \code{rad} object.
}
\examples{
data(BCI)
mod <- rad.lognormal(BCI[5,])

df <- fortify(mod$y)
ggplot(df, aes(Rank, Abundance)) +
    geom_point() +
    scale_y_log10()
   
}
\author{
Didzis Elferts
}
