% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.radline.R
\name{fortify.radline}
\alias{fortify.radline}
\title{Fortify \code{radline} object}
\usage{
\method{fortify}{radline}(model, data, ...)
}
\arguments{
\item{model}{an object of class \code{radline}.}

\item{data}{original data set. Currently ignored.}

\item{...}{additional arguments for other methods. Currently ignored.}
}
\value{
A data frame with columns \code{Species}, \code{Rank}, \code{Type} and \code{Abundance}, showing species name, their rank, type of values (observed or model fitted) and abundance values.
}
\description{
Prepares a fortified version of \code{radline} object.
}
\examples{
data(BCI)
mod <- rad.lognormal(BCI[5,])

df <- fortify(mod)
ggplot() +
    geom_point(data = df[df$Type == "Observed",], aes(Rank, Abundance)) +
    geom_line(data = df[df$Type == "Fitted",], aes(Rank, Abundance)) +
    scale_y_log10() + labs(color = "Model")
   
}
\author{
Didzis Elferts
}
