% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.rda.R
\name{fortify.rda}
\alias{fortify.rda}
\alias{tidy.rda}
\title{Fortify a \code{"rda"} object.}
\usage{
\method{fortify}{rda}(
  model,
  data = NULL,
  axes = 1:6,
  display = c("sp", "wa", "lc", "bp", "cn"),
  const = NULL,
  ...
)

\method{tidy}{rda}(
  x,
  data = NULL,
  axes = 1:6,
  display = c("sp", "wa", "lc", "bp", "cn"),
  const = NULL,
  ...
)
}
\arguments{
\item{model, x}{an object of class \code{"rda"}, the result of a call to
\code{\link[vegan:cca]{vegan::rda()}}.}

\item{data}{currently ignored.}

\item{axes}{numeric; which axes to extract scores for.}

\item{display}{numeric; the scores to extract in the fortified object.}

\item{const}{NULL; General scaling constant to RDA scores. See
\code{\link[vegan:plot.cca]{vegan::scores.rda()}} for the details.}

\item{...}{additional arguments passed to \code{\link[vegan:plot.cca]{vegan::scores.rda()}}.}
}
\value{
A data frame (tibble) in long format containing the ordination
scores. The first two components are \code{score} (the type of score in each
row) and \code{label} (the text label to use on plots for this row). The
remaining columns are the extracted ordination axis scores.
}
\description{
Fortifies an object of class \code{"rda"} to produce a
data frame of the selected axis scores in long format, suitable for
plotting with \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\details{
TODO
}
\examples{

library("vegan")

data(dune)
data(dune.env)

sol <- rda(dune ~ A1 + Management, data = dune.env)
head(fortify(sol))
}
\author{
Gavin L. Simpson
}
