% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.renyi.R
\name{fortify.renyi}
\alias{fortify.renyi}
\alias{tidy.renyi}
\title{Fortify vegan \code{renyi()} or \code{tsallis()} diversity results}
\usage{
\method{fortify}{renyi}(model, data, ...)

\method{tidy}{renyi}(x, data, ...)
}
\arguments{
\item{model}{model or other R object to convert to data frame}

\item{data}{original dataset, if needed}

\item{...}{Arguments passed to methods.}

\item{x}{An object to be converted into a tidy \code{\link[tibble:tibble]{tibble::tibble()}}.}
}
\value{
Returns a data frame or tibble in long format with columns
\code{diversity}, \code{scale}, \code{site}, \code{median}, \code{lo}, \code{hi}.
}
\description{
Prepares a fortified version of result objects from
\code{\link[vegan:renyi]{vegan::renyi()}} and \code{\link[vegan:tsallis]{vegan::tsallis()}}.
}
\author{
Jari Oksanen
}
