% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.wcmdscale.R
\name{fortify.wcmdscale}
\alias{fortify.wcmdscale}
\title{Fortify a \code{"wcmdscale"} object.}
\usage{
\method{fortify}{wcmdscale}(model, data, ...)
}
\arguments{
\item{model}{an object of class \code{"wcmdscale"}, the result of a call
to \code{\link[vegan]{wcmdscale}} with eig = TRUE.}

\item{data}{currently ignored.}

\item{...}{currently ignored.}
}
\value{
A data frame in long format containing the ordination scores.
}
\description{
Fortifies an object of class \code{"wcmdscale"} to produce a
data frame of the selected axis scores in long format, suitable for
plotting with \code{\link[ggplot2]{ggplot}}.
}
\details{
TODO
}
\examples{
## load example data
data(dune)

data(dune)
d <- vegdist(decostand(dune, method = 'pa'), method = 'jaccard')
ord <- wcmdscale(d, eig = TRUE)
head(fortify(ord))
}
\author{
Eduard Szöcs eduardszoecs@gmail.com
}

