% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiggplot.R
\name{geom_envfit}
\alias{geom_envfit}
\title{Add envfit Results to Ordination}
\usage{
geom_envfit(
  data,
  text = TRUE,
  box = FALSE,
  arrow.mul = 1,
  arrow.params = list(),
  text.params = list(),
  ...
)
}
\arguments{
\item{data}{\code{\link[vegan:envfit]{vegan::envfit()}} result object}

\item{text}{add text to plot.}

\item{box}{write text on a non-transparent label.}

\item{arrow.mul}{arro multiplier.}

\item{arrow.params, text.params}{List of additional parameters to
arrows and text.}

\item{...}{Other parameters passed to all graphical functions
\link{geom_ordi_arrow}, \link{geom_ordi_text} and \link{geom_ordi_label}.}
}
\description{
Function adds fitted vector arrays via \link{geom_ordi_arrow} and factor
centroids via \link{geom_ordi_text} or \link{geom_ordi_label}.
}
