% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiggplot.R
\name{geom_ordi_arrow}
\alias{geom_ordi_arrow}
\title{Add a biplot arrow layer to an ordiggplot}
\usage{
geom_ordi_arrow(
  score,
  data,
  text = TRUE,
  box = FALSE,
  arrow.mul = 1,
  arrow.params = list(),
  text.params = list(),
  ...
)
}
\arguments{
\item{score}{Ordination score to be added to the plot.}

\item{data}{Alternative data to the function that will be used
instead of \code{score}. This can be a \link[vegan:envfit]{vegan::envfit} result object
which is used to draw arrows to fitted vectors.}

\item{text}{Add text labels to the plot.}

\item{box}{Draw a box behind the text (logical).}

\item{arrow.mul}{Arrow multiplier.}

\item{arrow.params, text.params}{Parameters to modify arrows or
their text labels.}

\item{...}{other arguments passed to \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}},
\code{\link[ggplot2:geom_text]{ggplot2::geom_label()}}, or \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}}
}
\value{
Returns a ggplot2 layer or a list of such layers: a \code{"LayerInstance"}
object that inherits from classes \code{"Layer"}, \code{"ggproto"}, and \code{"gg"}.
}
\description{
Add a biplot arrow layer to an ordiggplot
}
