% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dimension_names}
\alias{get_dimension_names}
\title{Extract the names of the dimensions to plot as a character vector}
\usage{
get_dimension_names(object)
}
\arguments{
\item{object}{a fortified ordination object.}
}
\value{
A length 2 character vector of dimension names.
}
\description{
Find the character vector of names for the two dimensions of
data to be plotted.
}
\author{
Gavin L. Simpson
}
