% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_fun}
\alias{label_fun}
\title{Adds a label layer using one of a set of common geoms}
\usage{
label_fun(data, geom = c("label", "text", "label_repel", "text_repel"), vars)
}
\arguments{
\item{data}{data frame; data set to use for the label layer. Must contain a
variable label containing the strings to use as labels.}

\item{geom}{character; which geom to use for labelling.}

\item{vars}{character; vector of names of variables to pass to the \code{x} and
\code{y} aesthetics of the chosen geom.}
}
\description{
Adds labels to a plot using one of \code{\link[ggplot2:geom_text]{ggplot2::geom_label()}},
\code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}, \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_label_repel()}} or
\code{\link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel()}}.
}
\author{
Gavin L. Simpson
}
