% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{layer_draw_list}
\alias{layer_draw_list}
\title{List of layers to draw for a given vegan object}
\usage{
layer_draw_list(valid, layers = NULL, available = NULL)
}
\arguments{
\item{valid}{character; vector of valid layer names}

\item{layers}{character; a vector of layer names for \code{object} that has
already been filtered for validity.}

\item{available}{character; what layers are actually available}
}
\description{
List of layers to draw for a given vegan object
}
