% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scores_length}
\alias{scores_length}
\title{Number of scores}
\usage{
scores_length(x)
}
\arguments{
\item{x}{The object whose number of scores is required.}
}
\value{
a numeric vector of length 1 with the number of scores.
}
\description{
Returns the number of scores returns in object `x``.
}
\author{
Gavin L. Simpson
}
