% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{scoresLength}
\alias{scoresLength}
\title{Number of scores}
\usage{
scoresLength(x)
}
\arguments{
\item{x}{The object whose number of scores is required.}
}
\value{
a numeric vector of length 1 with the number of scores.
}
\description{
Returns the number of scores returns in object \code{x}.
}
\author{
Gavin L. Simpson
}

