% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_layers}
\alias{valid_layers}
\alias{valid_layers.rda}
\alias{valid_layers.cca}
\title{Valid layers for vegan objects}
\usage{
valid_layers(object, ...)

\method{valid_layers}{rda}(object, ...)

\method{valid_layers}{cca}(object, ...)
}
\arguments{
\item{object}{An R object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Valid layers for vegan objects
}
