% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiggplot.R
\name{autolayer.envfit}
\alias{autolayer.envfit}
\title{Add envfit Results to Ordination}
\usage{
\method{autolayer}{envfit}(
  object,
  text = TRUE,
  box = FALSE,
  arrow.mul = 1,
  arrow.params = list(),
  text.params = list(),
  ...
)
}
\arguments{
\item{object}{\code{\link[vegan:envfit]{vegan::envfit()}} result object}

\item{text}{add text to plot.}

\item{box}{write text on a non-transparent label.}

\item{arrow.mul}{arro multiplier.}

\item{arrow.params, text.params}{List of additional parameters to
arrows and text.}

\item{...}{Other parameters passed to all graphical functions
\link{geom_ordi_arrow}, \link{geom_ordi_text} and \link{geom_ordi_label}.}
}
\description{
Function adds fitted vector arrays via \link{geom_ordi_arrow} and factor
centroids via \link{geom_ordi_text} or \link{geom_ordi_label}.
}
\examples{
library(vegan)
library(ggplot2)
data(mite, mite.env, package = "vegan")
mod <- cca(mite)
ef <- envfit(mod ~ Shrub+Topo+WatrCont+SubsDens, mite.env)
ordiggplot(mod, scaling="sites") +
  geom_ordi_axis() +
  geom_ordi_point("sites") +
  autolayer(ef, arrow.mul=1.3, col="navy", box=TRUE,
    text.params=list(mapping=aes(fontface="bold")))

}
