% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiggplot.R
\name{autolayer.envfit}
\alias{autolayer.envfit}
\title{Add envfit Results to Ordination}
\usage{
\method{autolayer}{envfit}(
  object,
  text = TRUE,
  box = FALSE,
  arrow.mul = 1,
  arrow.params = list(),
  text.params = list(),
  ...
)
}
\arguments{
\item{object}{\code{\link[vegan:envfit]{vegan::envfit()}} result object}

\item{text}{add text to plot.}

\item{box}{write text on a non-transparent label.}

\item{arrow.mul}{arrow multiplier.}

\item{arrow.params, text.params}{List of additional parameters to
arrows and text.}

\item{...}{Other parameters passed to all graphical functions
\code{\link[=geom_ordi_arrow]{geom_ordi_arrow()}}, \code{\link[=geom_ordi_text]{geom_ordi_text()}} and \code{\link[=geom_ordi_label]{geom_ordi_label()}}.}
}
\value{
Returns ggplot2 layers \code{geom_segment} for arrows (when
appropriate), and \code{geom_text} or \code{geom_label} for text.
}
\description{
Function adds layer of fitted vector arrays and factor
centroids of factor levels in an \code{\link[=ordiggplot]{ordiggplot()}} graph.
}
\examples{
library(vegan)
library(ggplot2)
data(mite, mite.env, package = "vegan")
mod <- cca(mite)
## you must have same scaling in envfit() and ordiggplot()
ef <- envfit(mod ~ Shrub+Topo+WatrCont+SubsDens, mite.env,
  scaling = "sites")
ordiggplot(mod, scaling="sites") +
  geom_ordi_axis() +
  geom_ordi_point("sites") +
  autolayer(ef, arrow.mul=1.3, col="navy", box=TRUE,
    text.params=list(mapping=aes(fontface="bold")))

}
\author{
Jari Oksanen
}
