% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiggplot.R
\name{autolayer.ordisurf}
\alias{autolayer.ordisurf}
\title{Add ordisurf Result as Contours in Ordination plot}
\usage{
\method{autolayer}{ordisurf}(object, fill = FALSE, ...)
}
\arguments{
\item{object}{\link[vegan:ordisurf]{vegan::ordisurf} result object.}

\item{fill}{Use filled contours.}

\item{...}{Other arguments passed to \link[ggplot2:geom_contour]{ggplot2::geom_contour} or
\link[ggplot2:geom_contour]{ggplot2::geom_contour_filled}.}
}
\value{
A ggplot() layer.
}
\description{
Function adds \link[vegan:ordisurf]{vegan::ordisurf} result object
}
\examples{
library(vegan)
library(ggplot2)
data(mite, mite.env, package="vegan")
mod <- cca(mite)
surf <- ordisurf(mod ~ WatrCont, mite.env, plot = FALSE)
ordiggplot(mod) +
  geom_ordi_point("sites") +
  autolayer(surf) +
  autolayer(envfit(mod ~ WatrCont, mite.env, permutations=0),
    arrow.mul = 1.3)

}
\author{
Jari Oksanen
}
