% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.metaMDS.R
\name{autoplot.metaMDS}
\alias{autoplot.metaMDS}
\title{ggplot-based plot for objects of class \code{"metaMDS"}}
\usage{
\method{autoplot}{metaMDS}(object, axes = c(1, 2), geom = c("point",
  "text"), layers = c("species", "sites"), legend.position = "right",
  title = NULL, subtitle = NULL, caption = NULL, ylab, xlab, ...)
}
\arguments{
\item{object}{an object of class \code{"metaMDS"}, the result of a call
to \code{\link[vegan]{metaMDS}}.}

\item{axes}{numeric; which axes to plot, given as a vector of length 2.}

\item{geom}{character; which geoms to use for the layers. Can be a
vector of length equal to \code{length(display)}, in which case the
\emph{i}th element of \code{type} refers to the \emph{i}th element
of \code{display}.}

\item{layers}{character; which scores to plot as layers}

\item{legend.position}{character or two-element numeric vector; where to position the legend. See \code{\link[ggplot2]{theme}} for details. Use \code{"none"} to not draw the legend.}

\item{title}{character; subtitle for the plot}

\item{subtitle}{character; subtitle for the plot}

\item{caption}{character; caption for the plot}

\item{ylab}{character; label for the y-axis}

\item{xlab}{character; label for the x-axis}

\item{...}{Additional arguments passed to \code{\link{fortify.metaMDS}}.}
}
\value{
Returns a ggplot object.
}
\description{
Produces a multi-layer ggplot object representing the output of
objects produced by \code{\link[vegan]{metaMDS}}.
}
\details{
TODO
}
\examples{
## load example data
data(dune)

sol <- metaMDS(dune)
autoplot(sol)
}
\author{
Gavin L. Simpson
}
