% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.poolaccum.R
\name{autoplot.poolaccum}
\alias{autoplot.poolaccum}
\title{ggplot-based plot for objects of class \code{"poolaccum"}}
\usage{
\method{autoplot}{poolaccum}(object, alpha = 0.05, col = c("red", "blue"),
  nrow = 2, xlab = "Size", ylab = "Richness",
  title = "Accumulated species richness in a species pool", subtitle = NULL,
  caption = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{"poolaccum"}, the result of a call to \code{\link[vegan]{poolaccum}}.}

\item{alpha}{numeric; level of quantiles for envelopes shown (default 0.05).}

\item{col}{character; colors of mean Richness line and qunatiles envelope.}

\item{nrow}{numeric; number of rows in faceted plot.}

\item{xlab}{character; label for the x axis.}

\item{ylab}{character; label for the y axis.}

\item{title}{character; title for the plot.}

\item{subtitle}{character; subtitle for the plot.}

\item{caption}{character; caption for the plot.}

\item{...}{additional arguments passed to other methods.}
}
\value{
A ggplot object.
}
\description{
Produces a multi-layer ggplot object representing the output of objects produced by \code{\link[vegan]{poolaccum}}.
}
\examples{

data(BCI)
pool <- poolaccum(BCI)

autoplot(pool)
}
\author{
Didzis Elferts
}
