% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.radfit.R
\name{autoplot.radfit}
\alias{autoplot.radfit}
\title{ggplot-based plot for objects of class \code{"radfit"}}
\usage{
\method{autoplot}{radfit}(object, log = "y", leg.title = "Model",
  xlab = "Rank", ylab = "Abundance",
  title = "Species abundance against species rank", subtitle = NULL,
  caption = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{\link[vegan]{radfit}}.}

\item{log}{character; use logarithmic scale for "x" or "y", "xy" makes both axis logarthmic and "none" use original scale.}

\item{leg.title}{character; title of the legend.}

\item{xlab}{character; label for the x axis.}

\item{ylab}{character; label for the y axis.}

\item{title}{character; title for the plot.}

\item{subtitle}{character; subtitle for the plot.}

\item{caption}{character; caption for the plot.}

\item{...}{additional arguments passed to other methods.}
}
\value{
A ggplot object.
}
\description{
Produces a multi-layer ggplot object representing the output of objects produced by \code{\link[vegan]{radfit}}.
}
\examples{

data(BCI)
mod <- radfit(BCI[5,])

autoplot(mod)
}
\author{
Didzis Elferts
}
