% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.radfit.R
\name{autoplot.radfit}
\alias{autoplot.radfit}
\alias{fortify.radfit}
\title{Autoplot Graphics for vegan RAD models}
\usage{
\method{autoplot}{radfit}(object, facet = TRUE, ...)

\method{fortify}{radfit}(model, data, pick = NULL, ...)
}
\arguments{
\item{object}{Result object from \code{\link[vegan]{radfit}}.}

\item{facet}{Draw each fitted model to a separate facet or (if
\code{FALSE}) all fitted lines to a single graph.}

\item{...}{other arguments passed to methods}

\item{model}{model or other R object to convert to data frame}

\item{data}{original dataset, if needed}

\item{pick}{Pick or several models. Allowed values are \code{"AIC"}
and \code{"BIC"} for selecting the best model by AIC or BIC, or
(a vector of) model names that can be abbreviated. The default
returns all fitted models.}
}
\description{
\CRANpkg{ggplot2} graphics for Rank-Abundance Distribution models
fitted with \CRANpkg{vegan} function \code{\link[vegan]{radfit}}.
}
\details{
The \code{autoplot} function draws graphics which are \pkg{ggplot2}
alternatives for \CRANpkg{lattice} graphics in \pkg{vegan}. The
\code{fortify} function produces \dQuote{tidy} data frames that can
be used to produce the graphics.
}
\examples{
data(mite)
m1 <- radfit(mite[1,])
## With logarithmic y-axis (default) Pre-emption model is a line
autoplot(m1)
## With log-log scale, Zipf model is a line
autoplot(m1) + scale_x_log10()
}
