% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.renyi.R
\name{autoplot.renyi}
\alias{autoplot.renyi}
\alias{fortify.renyi}
\title{Autoplot Graphics for vegan Rényi and Tsallis diversity Objects}
\usage{
\method{autoplot}{renyi}(
  object,
  point.params = list(),
  line.params = list(),
  ribbon.params = list(),
  ...
)

\method{fortify}{renyi}(model, data, ...)
}
\arguments{
\item{object}{Result from \code{\link[vegan:renyi]{vegan::renyi()}} or \code{\link[vegan:tsallis]{vegan::tsallis()}}
functions.}

\item{point.params, line.params, ribbon.params}{List of graphical
parameters passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}},
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} or \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}}.}

\item{...}{Arguments passed to methods.}

\item{model}{model or other R object to convert to data frame}

\item{data}{original dataset, if needed}
}
\description{
\code{autoplot} for \code{\link[vegan:renyi]{vegan::renyi()}} and \code{\link[vegan:tsallis]{vegan::tsallis()}} plots
the diversites against diversity salce as dots over a ribbon of
data extremes in the graph and line of data median, faceted by
sites.
}
\examples{
## Rényi
library(vegan)
data(BCI)
mod <- renyi(BCI[sample(50, 16),])
autoplot(mod)

}
