% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.renyi.R
\name{autoplot.renyi}
\alias{autoplot.renyi}
\title{Autoplot Graphics for vegan \code{renyi()} and \code{tsallis()} diversity
Objects}
\usage{
\method{autoplot}{renyi}(
  object,
  point.params = list(),
  line.params = list(),
  ribbon.params = list(),
  ...
)
}
\arguments{
\item{object}{an object, whose class will determine the behaviour of autoplot}

\item{point.params, line.params, ribbon.params}{List of graphical
parameters passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}},
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} or \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}}.}

\item{...}{other arguments passed to specific methods}
}
\description{
\code{autoplot} for \code{\link[vegan:renyi]{vegan::renyi()}} and \code{\link[vegan:tsallis]{vegan::tsallis()}} plots
the diversities against diversity scale as dots separately for each
site. As a background it uses a ribbon of data extremes and a line of
data median in all sites.
}
\examples{
## Rényi
library(vegan)
data(BCI)
mod <- renyi(BCI[1:16,])
autoplot(mod)

}
