% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.renyi.R
\name{autoplot.renyi}
\alias{autoplot.renyi}
\title{ggplot-based plot for objects of class \code{"renyi"}}
\usage{
\method{autoplot}{renyi}(object, facet = TRUE, ribbon = TRUE, ncol = NULL,
  ribbon.alpha = 0.3, xlab = "Scale", ylab = "Diversity",
  title = "Diversity at different scales", subtitle = NULL,
  caption = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{renyi}, the result of a call to \code{\link[vegan]{renyi}}.}

\item{facet}{logical; should diversity values be shown in separate facets if more than one site present?}

\item{ribbon}{logical; show the quantile-based uncertainty interval? Uses \code{\link{geom_ribbon}} for plot.}

\item{ncol}{numeric; if facetting the plot, how many columns to use.}

\item{ribbon.alpha}{numeric; alpha transparency used for the uncertainty interval. Passed to the \code{alpha} aesthetic of \code{\link[ggplot2]{geom_ribbon}}.}

\item{xlab}{character; label for the x axis.}

\item{ylab}{character; label for the y axis.}

\item{title}{character; title for the plot.}

\item{subtitle}{character; subtitle for the plot.}

\item{caption}{character; caption for the plot.}

\item{...}{additional arguments passed to other methods.}
}
\value{
A ggplot object.
}
\description{
Produces a multi-layer ggplot object representing the output of objects produced by \code{\link[vegan]{renyi}}.
}
\examples{

data(BCI)
i <- sample(nrow(BCI), 12)
mod <- renyi(BCI[i,])

autoplot(mod)
}
\author{
Didzis Elferts & Gavin L. Simpson
}
