% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.renyi.R
\name{autoplot.renyi}
\alias{autoplot.renyi}
\title{ggplot-based plot for objects of class \code{"renyi"}}
\usage{
\method{autoplot}{renyi}(object, col = c("red", "blue"), nrow = 3,
  xlab = "Scale", ylab = "Diversity",
  title = "Diversity at different scales", subtitle = NULL,
  caption = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{renyi}, the result of a call to \code{\link[vegan]{renyi}}.}

\item{col}{character; colors of median diversity line and qunatiles envelope.}

\item{nrow}{numeric; number of rows in faceted plot.}

\item{xlab}{character; label for the x axis.}

\item{ylab}{character; label for the y axis.}

\item{title}{character; title for the plot.}

\item{subtitle}{character; subtitle for the plot.}

\item{caption}{character; caption for the plot.}

\item{...}{additional arguments passed to other methods.}
}
\value{
A ggplot object.
}
\description{
Produces a multi-layer ggplot object representing the output of objects produced by \code{\link[vegan]{renyi}}.
}
\examples{

data(BCI)
i <- sample(nrow(BCI), 12)
mod <- renyi(BCI[i,])

autoplot(mod)
}
\author{
Didzis Elferts
}
