% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.renyiaccum.R
\name{autoplot.renyiaccum}
\alias{autoplot.renyiaccum}
\title{ggplot-based plot for objects of class \code{"renyiaccum"}}
\usage{
\method{autoplot}{renyiaccum}(object, col = c("red", "blue"), nrow = 2,
  xlab = "Sites", ylab = "Diversity",
  title = "Diversity accumulation curves", subtitle = NULL,
  caption = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{renyiaccum}, the result of a call to \code{\link[vegan]{renyiaccum}}.}

\item{col}{character; colors of mean diversity line and qunatiles envelope.}

\item{nrow}{numeric; number of rows in faceted plot.}

\item{xlab}{character; label for the x axis.}

\item{ylab}{character; label for the y axis.}

\item{title}{character; title for the plot.}

\item{subtitle}{character; subtitle for the plot.}

\item{caption}{character; caption for the plot.}

\item{...}{additional arguments passed to other methods.}
}
\value{
A ggplot object.
}
\description{
Produces a multi-layer ggplot object representing the output of objects produced by \code{\link[vegan]{renyiaccum}}.
}
\examples{

data(BCI)
i <- sample(nrow(BCI), 12)
mod <- renyiaccum(BCI[i,])

autoplot(mod)
}
\author{
Didzis Elferts
}
