% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.anosim.R
\name{fortify.anosim}
\alias{fortify.anosim}
\title{Fortify analysis of similarities (ANOSIM) results}
\usage{
\method{fortify}{anosim}(model, data, ...)
}
\arguments{
\item{model}{an object of class \code{\link[vegan]{anosim}}.}

\item{data}{original data set. Currently ignored.}

\item{...}{additional arguments for other methods. Currently ignored.}
}
\value{
A data frame with columns \code{'Rank'} and \code{'Class'} with ranks of dissimilarity.
}
\description{
Prepares fortified versions of results from \code{\link[vegan]{anosim}} objects.
}
\examples{
data(dune)
data(dune.env)
dune.dist <- vegdist(dune)
dune.ano <- with(dune.env, anosim(dune.dist, Management))

df <- fortify(dune.ano)

ggplot(df, aes(x = Class, y = Rank)) +
      geom_boxplot(notch = TRUE, varwidth = TRUE) + 
      theme(axis.title = element_blank())
      
}
\author{
Didzis Elferts
}
