% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.cca.R
\name{fortify.cca}
\alias{fortify.cca}
\title{Fortify a \code{"cca"} object.}
\usage{
\method{fortify}{cca}(
  model,
  data,
  axes = 1:6,
  display = c("sp", "wa", "lc", "bp", "cn"),
  ...
)
}
\arguments{
\item{model}{an object of class \code{"cca"}, the result of a call to
\code{\link[vegan:cca]{vegan::cca()}}, \code{\link[vegan:cca]{vegan::rda()}}, or \code{\link[vegan:dbrda]{vegan::capscale()}}.}

\item{data}{currently ignored.}

\item{axes}{numeric; which axes to extract scores for.}

\item{display}{numeric; the scores to extract in the fortified object.}

\item{...}{additional arguments passed to \code{\link[vegan:plot.cca]{vegan::scores.cca()}},
and \code{\link[vegan:plot.cca]{vegan::scores.rda()}}.}
}
\value{
A data frame in long format containing the ordination scores.
The first two components are the axis scores.
}
\description{
Fortifies an object of class \code{"cca"} to produce a
data frame of the selected axis scores in long format, suitable for
plotting with \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\details{
TODO
}
\examples{

library("vegan")

data(dune)
data(dune.env)

sol <- cca(dune ~ A1 + Management, data = dune.env)
head(fortify(sol))
}
\author{
Gavin L. Simpson
}
