% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.metaMDS.R
\name{fortify.metaMDS}
\alias{fortify.metaMDS}
\title{Fortify a \code{"mataMDS"} object.}
\usage{
\method{fortify}{metaMDS}(model, data, ...)
}
\arguments{
\item{model}{an object of class \code{"metaMDS"}, the result of a call
to \code{\link[vegan]{metaMDS}}.}

\item{data}{currently ignored.}

\item{...}{additional arguments passed to
\code{\link[vegan]{scores.metaMDS}}. Note you can't use \code{display}.}
}
\value{
A data frame in long format containing the ordination scores.
The first two components are the axis scores.
}
\description{
Fortifies an object of class \code{"metaMDS"} to produce a
data frame of the selected axis scores in long format, suitable for
plotting with \code{\link[ggplot2]{ggplot}}.
}
\details{
TODO
}
\examples{

library("vegan")

data(dune)

ord <- metaMDS(dune)
head(fortify(ord))
}
\author{
Gavin L. Simpson
}
