% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.poolaccum.R
\name{fortify.poolaccum}
\alias{fortify.poolaccum}
\title{Fortify accumulated species richness in as species pool results}
\usage{
\method{fortify}{poolaccum}(model, data, alpha = 0.05, ...)
}
\arguments{
\item{model}{an object of class \code{\link[vegan]{poolaccum}}.}

\item{data}{original data set. Currently ignored.}

\item{alpha}{level of quantiles for envelopes shown (default 0.05).}

\item{...}{additional arguments for other methods. Currently ignored.}
}
\value{
A data frame with columns \code{Size}, \code{Richness}, \code{Lower}, \code{Upper}, \code{Std.dev} and \code{Index}, showing permutation richness estimator, sample size, standart deviation, quantile envelope and indices names.
}
\description{
Prepares a fortified version of results from \code{\link[vegan]{poolaccum}} objects.
}
\examples{
data(BCI)
pool <- poolaccum(BCI)
df <- fortify(pool)

ggplot(df, aes(x = Size)) + 
   geom_line(aes(y = Richness), color = "red") +
   geom_line(aes(y = Lower), color = "blue") +
   geom_line(aes(y = Upper), color = "blue") +
   facet_wrap(~Index, ncol = 3) 
}
\author{
Didzis Elferts
}
