% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.prestonfit.R
\name{fortify.prestonfit}
\alias{fortify.prestonfit}
\title{Tidied species octave abundance data}
\usage{
\method{fortify}{prestonfit}(model, data, ...)
}
\arguments{
\item{model}{an object of class \code{\link{prestonfit}}}

\item{data}{original data set. Currently ignored.}

\item{...}{other arguments pass to methods. Currently ignored.}
}
\value{
A data frame with columns \code{'Octave'} and \code{'Abundance'}.
}
\description{
Prepares a data frame of results from a \code{\link{fisherfit}} object suitable for plotting with \code{\link{ggplot}}.
}
\examples{

data(BCI)
pfit <- prestonfit(colSums(BCI))
head(fortify(pfit))
}
\author{
Gavin L. Simpson
}
