% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.radfit.R
\name{fortify.radfit}
\alias{fortify.radfit}
\alias{fortify.radfit.frame}
\alias{fortify.radline}
\alias{fortify.rad}
\alias{fortify.rad.frame}
\alias{tidy.radfit}
\alias{tidy.radfit.frame}
\alias{tidy.radline}
\alias{tidy.rad}
\alias{tidy.rad.frame}
\title{Fortify results of ranked abundance distribution models}
\usage{
\method{fortify}{radfit}(model, data, pick = NULL, ...)

\method{fortify}{radfit.frame}(model, data, pick = "AIC", order.by = NULL, ...)

\method{fortify}{radline}(model, data, ...)

\method{fortify}{rad}(model, data, ...)

\method{fortify}{rad.frame}(model, data, order.by = NULL, ...)

\method{tidy}{radfit}(x, data, pick = NULL, ...)

\method{tidy}{radfit.frame}(x, data, pick = "AIC", order.by = NULL, ...)

\method{tidy}{radline}(x, data, ...)

\method{tidy}{rad}(x, data, ...)

\method{tidy}{rad.frame}(x, data, order.by = NULL, ...)
}
\arguments{
\item{model}{model or other R object to convert to data frame}

\item{data}{original dataset, if needed}

\item{pick}{Pick or several models. Allowed values are \code{"AIC"}
and \code{"BIC"} for selecting the best model by AIC or BIC, or
(a vector of) model names that can be abbreviated. The default
returns all fitted models.}

\item{...}{Arguments passed to methods.}

\item{order.by}{A vector used for ordering site panels.}

\item{x}{An object to be converted into a tidy \code{\link[tibble:tibble]{tibble::tibble()}}.}
}
\value{
A data frame or tibble which in all cases has columns
\code{species} (character), for species names, with \code{rank} and
\code{abundance} for data, and may also have columns \code{site} (factor)
for site names when several sites were analysed, \code{fit} of
fitted abundances when evaluated, and \code{model} (factor) when
several models were fitted.
}
\description{
Prepares a fortified version of results from
\code{\link[vegan:radfit]{vegan::radfit()}} or \code{\link[vegan:radfit]{vegan::as.rad()}}.
}
\author{
Jari Oksanen
}
