% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.radfit.R
\name{fortify.radfit}
\alias{fortify.radfit}
\title{Fortify \code{radfit} object}
\usage{
\method{fortify}{radfit}(model, data, ...)
}
\arguments{
\item{model}{an object of class \code{radfit}.}

\item{data}{original data set. Currently ignored.}

\item{...}{additional arguments for other methods. Currently ignored.}
}
\value{
A data frame with columns \code{Species}, \code{Rank}, \code{Type} and \code{Abundance}, showing species name, their rank, type of values (observed or model fitted) and abundance values.
}
\description{
Prepares a fortified version of \code{\link[vegan]{radfit}} object.
}
\examples{
data(BCI)
mod <- radfit(BCI[5,])

df <- fortify(mod)
ggplot() +
    geom_point(data = df[df$Type == "Observed",], aes(Rank, Abundance)) +
    geom_line(data = df[df$Type != "Observed",], aes(Rank, Abundance, color = Type)) +
    scale_y_log10() + labs(color = "Model")
   
}
\author{
Didzis Elferts
}
