% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.renyi.R
\name{fortify.renyi}
\alias{fortify.renyi}
\title{Fortify Renyi and Hill diversity objects}
\usage{
\method{fortify}{renyi}(model, data, ...)
}
\arguments{
\item{model}{an object of class \code{\link[vegan]{renyi}}.}

\item{data}{original data set. Currently ignored.}

\item{...}{additional arguments for other methods. Currently ignored.}
}
\value{
A data frame with columns \code{Scale}, \code{Diversity} and \code{Site}, showing divesity values at different scales. Column \code{Site} produced if more than one site used in analysis.
}
\description{
Prepares a fortified version from \code{\link[vegan]{renyi}} objects.
}
\examples{
data(BCI)
i <- sample(nrow(BCI), 12)
mod <- renyi(BCI[i,])
df <- fortify(mod)

ggplot(df, aes(x = Scale, y = Diversity)) + 
   geom_point() +
   facet_wrap(~Site, ncol = 3) 
   
}
\author{
Didzis Elferts
}
