% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.renyiaccum.R
\name{fortify.renyiaccum}
\alias{fortify.renyiaccum}
\title{Fortify Renyi and Hill diversity accumulation objects}
\usage{
\method{fortify}{renyiaccum}(model, data, ...)
}
\arguments{
\item{model}{an object of class \code{\link[vegan]{renyiaccum}}.}

\item{data}{original data set. Currently ignored.}

\item{...}{additional arguments for other methods. Currently ignored.}
}
\value{
A data frame with columns \code{Sites}, \code{Scale}, \code{Lower}, \code{Max}, \code{Mean}, \code{Min}, \code{St.dev} and \code{Upper}, showing divesity values at different scales for different number of sites.
}
\description{
Prepares a fortified version from \code{\link[vegan]{renyiaccum}} objects.
}
\examples{
data(BCI)
i <- sample(nrow(BCI), 12)
mod <- renyiaccum(BCI[i,])
df <- fortify(mod)

ggplot(df, aes(Sites, Mean)) +
  geom_line() + 
  facet_wrap(~ Scale, nrow = 2)  
   
}
\author{
Didzis Elferts
}
