% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.renyiaccum.R
\name{fortify.renyiaccum}
\alias{fortify.renyiaccum}
\title{Fortify Renyi and Hill diversity accumulation objects}
\usage{
\method{fortify}{renyiaccum}(model, data, ...)
}
\arguments{
\item{model}{an object of class \code{\link[vegan]{renyiaccum}}.}

\item{data}{original data set. Currently ignored.}

\item{...}{additional arguments for other methods. Currently ignored.}
}
\value{
A data frame with columns \code{Scale}, \code{Diversity}, \code{Sites} and \code{Parameter}, showing divesity values at different scales for different number of sites. Column \code{Parameter} contains type of diversity value - mean, standart deviation, minimal, maximal, 2.5% and 97.5% quantiles.
}
\description{
Prepares a fortified version from \code{\link[vegan]{renyiaccum}} objects.
}
\examples{
data(BCI)
i <- sample(nrow(BCI), 12)
mod <- renyiaccum(BCI[i,])
df <- fortify(mod)

ggplot() +
  geom_line(data = df[df$Parameter == "mean",], aes(Sites, Diversity), color = "red") +
  geom_line(data = df[df$Parameter == "Qnt 0.025",], aes(Sites, Diversity), color = "blue") +
  geom_line(data = df[df$Parameter == "Qnt 0.975",], aes(Sites, Diversity), color = "blue") +
  facet_wrap(~ Scale, nrow = 2)  
   
}
\author{
Didzis Elferts
}
