% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiggplot.R
\name{geom_ordi_arrow}
\alias{geom_ordi_arrow}
\title{Add an Arrow Layer to an ordiggplot Graph.}
\usage{
geom_ordi_arrow(
  score,
  data,
  text = TRUE,
  box = FALSE,
  arrow.mul = 1,
  arrow.params = list(),
  text.params = list(),
  ...
)
}
\arguments{
\item{score}{Ordination score to be added to the plot.}

\item{data}{Alternative data to the function that will be used
instead of \code{score}. This can be a \link[vegan:envfit]{vegan::envfit} result object
which is used to draw arrows of fitted vectors.}

\item{text}{Add text labels to the plot (logical).}

\item{box}{Draw a non-transparent box behind the text (logical).}

\item{arrow.mul}{Arrow multiplier when arrows end points are given
in \code{data}. Ignored for \code{biplot} and \code{regression} scores or
\code{species} scores which are already scaled in constrained
ordination methods when calling \code{\link[=ordiggplot]{ordiggplot()}}.}

\item{arrow.params, text.params}{Parameters to modify arrows or
their text labels.}

\item{...}{other arguments passed to \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}},
\code{\link[ggplot2:geom_text]{ggplot2::geom_label()}} and \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}}
}
\value{
Returns ggplot2 layers \code{geom_segment} for arrows, and
\code{geom_text} or \code{geom_label} (optionally) for their names.
}
\description{
Function adds layers of arrows (\code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}) and
(optionally) their text labels (\code{\link[ggplot2:geom_text]{ggplot2::geom_text()}} or
\code{\link[ggplot2:geom_text]{ggplot2::geom_label()}}) to an \code{\link[=ordiggplot]{ordiggplot()}} graph. Typically
these arrows are \code{biplot} or \code{regression} scores, but the function
allows drawing species arrows for PCA biplots, or adding fitted
environmental vectors of \code{\link[vegan:envfit]{vegan::envfit()}}.
}
\examples{
library(vegan)
data(varespec, varechem, package="vegan")
## PCA biplot
mod <- rda(varespec)
ordiggplot(mod, scaling = "sites", corr = TRUE) +
  geom_ordi_axis() +
  geom_ordi_point("sites") +
  geom_ordi_arrow("species")
## RDA
mod <- rda(varespec ~ N + K + Al + Baresoil, varechem)
ordiggplot(mod) +
  geom_ordi_axis() +
  geom_ordi_point("sites") +
  geom_ordi_arrow("biplot", box = TRUE)

}
\seealso{
Underlying functions are \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}},
\code{\link[ggplot2:geom_text]{ggplot2::geom_text()}} and \code{\link[ggplot2:geom_text]{ggplot2::geom_label()}}.
}
\author{
Jari Oksanen
}
