% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiggplot.R
\name{geom_ordi_axis}
\alias{geom_ordi_axis}
\title{Crosshair for axes in eigenvector methods}
\usage{
geom_ordi_axis(lty = 3, ...)
}
\arguments{
\item{lty}{Linetype.}

\item{...}{other arguments passed to \code{\link[ggplot2:geom_abline]{ggplot2::geom_hline()}} and
\code{\link[ggplot2:geom_abline]{ggplot2::geom_vline()}}}
}
\description{
Crosshair for axes in eigenvector methods
}
