% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiggplot.R
\name{geom_ordi_label}
\alias{geom_ordi_label}
\title{Add a Label Layer to an ordiggplot Graph}
\usage{
geom_ordi_label(score, data, ...)
}
\arguments{
\item{score}{Ordination score to be added to the plot.}

\item{data}{Alternative data to the function that will be used
instead of \code{score}. This can be a \code{\link[vegan:envfit]{vegan::envfit()}} result object
which is used to plot centroids of factor levels.}

\item{...}{other arguments passed to \code{\link[ggplot2:geom_text]{ggplot2::geom_label()}}}
}
\value{
Returns a ggplot2 layer \code{geom_label}.
}
\description{
Function adds \dQuote{labels} or text in a box with non-transparent
background using \code{\link[ggplot2:geom_text]{ggplot2::geom_label()}} to an \code{\link[=ordiggplot]{ordiggplot()}}
graph. These can help in congested plot since uppermost labels are
readable (but cover lower ones), or to emphasize text. Some
\pkg{ggvegan} functions call \code{\link[ggplot2:geom_text]{ggplot2::geom_label()}} with argument
\code{box = TRUE}.
}
\examples{
library(vegan)
library(ggplot2)
data(dune, dune.env, package = "vegan")
mod <- cca(dune ~ Moisture, dune.env)
ordiggplot(mod) +
  geom_ordi_axis() +
  geom_ordi_point("sites") +
  geom_ordi_label("species", size = 3,
    mapping = aes(fontface = "italic")) +
  geom_ordi_label("centroids", size = 5,
    mapping = aes(fontface = "bold"))

}
\seealso{
\code{\link[ggplot2:geom_text]{ggplot2::geom_label()}}.
}
\author{
Jari Oksanen
}
