% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiggplot.R
\name{geom_ordi_label}
\alias{geom_ordi_label}
\title{Add a label layer to an ordiggplot}
\usage{
geom_ordi_label(score, data, ...)
}
\arguments{
\item{score}{Ordination score to be added to the plot.}

\item{data}{Alternative data to the function that will be used
instead of \code{score}.}

\item{...}{other arguments passed to \code{\link[ggplot2:geom_text]{ggplot2::geom_label()}}}
}
\value{
Returns a ggplot object.
}
\description{
Add a label layer to an ordiggplot
}
