% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiggplot.R
\name{geom_ordi_point}
\alias{geom_ordi_point}
\title{Add a point layer to an ordiggplot}
\usage{
geom_ordi_point(score, data, ...)
}
\arguments{
\item{score}{Ordination score to be added to the plot.}

\item{data}{Alternative data to the function that will be used
instead of \code{score}.}

\item{...}{other arguments passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}}
}
\value{
Returns a ggplot2 layer \code{geom_point}.
}
\description{
Add a point layer to an ordiggplot
}
