% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiggplot.R
\name{geom_ordi_repel}
\alias{geom_ordi_repel}
\title{Add points and their labels to the plot. The labels are repelled
from other pitems to minimize overplotting.}
\usage{
geom_ordi_repel(
  score,
  data,
  box = FALSE,
  points = TRUE,
  point.params = list(),
  text.params = list(),
  ...
)
}
\arguments{
\item{score}{Ordination score added to the plot.}

\item{data}{Alternative data to the function that will be used
instead of \code{score}. This function does \emph{not} handle
\code{\link[vegan:envfit]{vegan::envfit()}} results.}

\item{box}{Draw a box behind the text (logical).}

\item{points}{Draw points (logical).}

\item{point.params, text.params}{Parameters to modify points or
their repelled text labels.}

\item{...}{other arguments passed to \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel()}}
and \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_label_repel()}}.}
}
\value{
Returns ggrepel layers \code{geom_text_repel} or
\code{geom_label_repel} and  ggplot2 layer \code{geom_point} (optionally).
}
\description{
Function adds points to the exact position of the ordination score
with a label, but labels are repelled from each other to avoid
over-plotting. In very congested areas, the labels are completely
omitted.
}
\examples{
library(vegan)
data(mite, mite.env, package = "vegan")
mod <- cca(mite)
ordiggplot(mod) +
  geom_ordi_axis() +
  geom_ordi_point("sites") +
  geom_ordi_repel("species",
    text.params = list(size=3, fontface = "italic"))

}
