% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiggplot.R
\name{ordiggplot}
\alias{ordiggplot}
\alias{geom_ordipoint}
\alias{geom_orditext}
\alias{geom_ordiarrow}
\alias{ggscores}
\title{Create a ggplot Object}
\usage{
ordiggplot(model, axes = c(1, 2), ...)

geom_ordipoint(Score, ...)

geom_orditext(Score, box = FALSE, ...)

geom_ordiarrow(Score, text = TRUE, box = FALSE, ...)

ggscores(Score)
}
\arguments{
\item{model}{An ordination result object from \CRANpkg{vegan}.}

\item{axes}{Two axes to be plotted}

\item{...}{Parameters passed to underlying functions.}

\item{Score}{Ordination score to be added to the plot.}

\item{box}{Draw a box behind the text (logical).}

\item{text}{Add text labels to the plot.}
}
\description{
Function \code{ordiggplot} sets up an ordination graph but draws no
result. You can add new graphical elements to this plot with
\code{geom_ordi*} function of this package, or you can use standard
\CRANpkg{ggplot2} \code{geom_*} functions and use \code{ggscores}
as their \code{data=} argument.
}
\details{
The \pkg{ggvegan} package has two contrasting approaches to draw
ordination plots. The \code{autoplot} functions
(e.g. \code{\link{autoplot.rda}}, \code{\link{autoplot.cca}},
\code{\link{autoplot.metaMDS}}) draw a complete plot with one
command, but the design is hard-coded in the function. However, you
but can add new elements to the graph.

In contrast, function \code{ordiggplot} only sets up an ordination
plot, but draws no result. It allows you to add layers to the graph
one by one with full flexibility of the \CRANpkg{ggplot2}
functions. There are some specific functions \code{geom_ordi*}
functions that are similar as similarly named \code{geom_*}
functions. For these you need to give the name of ordination score
to be added, and in addition, you can give any \code{geom_}
function arguments to modify the plot. Alternatively, you can use
any \pkg{ggplot2} function and in its \code{data=} argument use
\code{ggscores} function to select the data elements for the
function.

The \code{ordiggplot} function extracts results using
\code{fortify} functions of this package, and it accepts the
arguments of those functions. This allows setting, e.g., the
scaling of ordination axes.
}
\examples{
data(dune, dune.env)
m <- cca(dune ~ Management, dune.env)
## use geom_ordi* functions
ordiggplot(m) + geom_ordipoint("sites") +
  geom_orditext("species", col="darkblue", mapping=aes(fontface="italic")) +
  geom_orditext("centroids", box=TRUE)
## use ggscores + standard geom_* functions
ordiggplot(m, scaling="sites") + geom_point(data=ggscores("sites")) +
   geom_text(data=ggscores("species"), mapping=aes(fontface="italic")) +
   geom_label(data=ggscores("centroids"), fill="yellow")


}
