% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiggplot.R
\name{ordiggplot}
\alias{ordiggplot}
\alias{geom_ordipoint}
\alias{geom_orditext}
\title{Create a ggplot Object}
\usage{
ordiggplot(model, axes = c(1, 2), ...)

geom_ordipoint(Score, ...)

geom_orditext(Score, box = FALSE, ...)
}
\arguments{
\item{model}{An ordination result object from \CRANpkg{vegan}.}

\item{axes}{Two axes to be plotted}

\item{...}{Parameters passed to underlying functions.}

\item{Score}{Ordination score to be added to the plot.}

\item{box}{Draw a box behind the text (logical).}
}
\description{
Function \code{ordiggplot} set ups an ordination plot. The function
extracts results using \code{fortify} functions of this package,
and it accepts the arguments of those functions. This allows
setting, e.g., the scaling of ordination axes. Functions
\code{geom_ordipoint} and \code{geom_orditext} add points or gext
to this plot. The \code{geom_ordi} functions accept arguments of
\code{\link[ggplot2]{geom_point}} and
\code{\link[ggplot2]{geom_text}} allowing flexible editing of
plots.
}
\examples{
data(dune, dune.env)
m <- cca(dune ~ Management, dune.env)
ordiggplot(m) + geom_ordipoint("sites") +
  geom_orditext("species", col="darkblue", mapping=aes(fontface="italic")) +
  geom_orditext("centroids", box=TRUE)

}
